requirements/base.txt
pytz==2018.5  # https://github.com/stub42/pytz
python-slugify==1.2.5  # https://github.com/un33k/python-slugify
Pillow==5.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.3.0  # https://github.com/hynek/argon2_cffi
redis>=2.10.5  # https://github.com/antirez/redis
celery==4.2.1  # pyup: <5.0  # https://github.com/celery/celery

# Django
# ------------------------------------------------------------------------------
django==2.0.8  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.37.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Stuff I Added
django-cors-headers==2.4.0
djangorestframework-jwt==1.11.0
django-rest-auth-0.9.3
